+++ src/main/java/de/maxhenkel/car/blocks/BlockFluidPipe.java
@@ -5,28 +5,30 @@ import de.maxhenkel.car.ModItemGroups;
 import de.maxhenkel.tools.VoxelShapeTools;
 import de.maxhenkel.tools.FluidUtils;
 import de.maxhenkel.tools.IItemBlock;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockRenderType;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.SoundType;
+import net.minecraft.block.*;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.material.MaterialColor;
+import net.minecraft.fluid.Fluids;
+import net.minecraft.fluid.IFluidState;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.BlockItemUseContext;
 import net.minecraft.item.Item;
 import net.minecraft.state.BooleanProperty;
 import net.minecraft.state.StateContainer;
+import net.minecraft.state.properties.BlockStateProperties;
+import net.minecraft.tags.FluidTags;
 import net.minecraft.util.Direction;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.world.IBlockReader;
+import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
 
 import javax.annotation.Nullable;
 
-public class BlockFluidPipe extends BlockBase implements IItemBlock {
+public class BlockFluidPipe extends BlockBase implements IItemBlock, IWaterLoggable {
 
     public static final BooleanProperty DOWN = BooleanProperty.create("down");
     public static final BooleanProperty UP = BooleanProperty.create("up");
@@ -34,6 +36,7 @@ public class BlockFluidPipe extends BlockBase implements IItemBlock {
     public static final BooleanProperty SOUTH = BooleanProperty.create("south");
     public static final BooleanProperty WEST = BooleanProperty.create("west");
     public static final BooleanProperty EAST = BooleanProperty.create("east");
+    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 
     public BlockFluidPipe() {
         super(Properties.create(Material.WOOL, MaterialColor.GRAY).hardnessAndResistance(0.25F).sound(SoundType.METAL));
@@ -46,6 +49,7 @@ public class BlockFluidPipe extends BlockBase implements IItemBlock {
                 .with(SOUTH, false)
                 .with(EAST, false)
                 .with(WEST, false)
+                .with(WATERLOGGED, false)
         );
     }
 
@@ -61,13 +65,23 @@ public class BlockFluidPipe extends BlockBase implements IItemBlock {
     }
 
     private BlockState getState(World world, BlockPos pos) {
+        IFluidState ifluidstate = world.getFluidState(pos);
         return getDefaultState()
                 .with(UP, isConnectedTo(world, pos, Direction.UP))
                 .with(DOWN, isConnectedTo(world, pos, Direction.DOWN))
                 .with(NORTH, isConnectedTo(world, pos, Direction.NORTH))
                 .with(SOUTH, isConnectedTo(world, pos, Direction.SOUTH))
                 .with(EAST, isConnectedTo(world, pos, Direction.EAST))
-                .with(WEST, isConnectedTo(world, pos, Direction.WEST));
+                .with(WEST, isConnectedTo(world, pos, Direction.WEST))
+                .with(WATERLOGGED, ifluidstate.isTagged(FluidTags.WATER) && ifluidstate.getLevel() == 8);
+    }
+
+    @Override
+    public BlockState updatePostPlacement(BlockState stateIn, Direction facing, BlockState facingState, IWorld worldIn, BlockPos currentPos, BlockPos facingPos) {
+        if (stateIn.get(WATERLOGGED)) {
+            worldIn.getPendingFluidTicks().scheduleTick(currentPos, Fluids.WATER, Fluids.WATER.getTickRate(worldIn));
+        }
+        return super.updatePostPlacement(stateIn, facing, facingState, worldIn, currentPos, facingPos);
     }
 
     @Override
@@ -78,7 +92,7 @@ public class BlockFluidPipe extends BlockBase implements IItemBlock {
 
     @Override
     protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
-        builder.add(UP, DOWN, NORTH, SOUTH, EAST, WEST);
+        builder.add(UP, DOWN, NORTH, SOUTH, EAST, WEST, WATERLOGGED);
     }
 
     public static final VoxelShape SHAPE_NORTH = Block.makeCuboidShape(6D, 6D, 6D, 10D, 10D, 0D);
@@ -127,6 +141,11 @@ public class BlockFluidPipe extends BlockBase implements IItemBlock {
         }
 
         return FluidUtils.isFluidHandler(world, pos, facing);
+    }
+
+    @Override
+    public IFluidState getFluidState(BlockState state) {
+        return state.get(WATERLOGGED) ? Fluids.WATER.getStillFluidState(false) : super.getFluidState(state);
     }
 
     @Override